修复产品获取功能中的 ID 类型，确保使用无符号整型，并优化数据库连接字符串格式

+++ app/product/biz/model/product.go
@@ -19,8 +19,8 @@ func (p Product) TableName() string {
 	return "product"
 }
 
-func GetProductById(ctx context.Context, db *gorm.DB, id int) (product Product, err error) {
-	err = db.WithContext(ctx).Model(&Product{}).Where(&Product{Model: gorm.Model{ID: uint(id)}}).First(&product).Error
+func GetProductById(ctx context.Context, db *gorm.DB, id uint) (product Product, err error) {
+	err = db.WithContext(ctx).Model(&Product{}).Where(&Product{Model: gorm.Model{ID: id}}).First(&product).Error
 	return
 }
